api/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.3
astroid==2.4.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-adminlte3==0.1.6
django-bootstrap3==14.1.0
django-cors-headers==3.7.0
django-group-by==0.3.1
django-shopping-cart==0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
future==0.18.2
html5lib==1.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient @ file:///C:/Users/Dev/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==7.1.2
PyJWT==1.7.1
pylint==2.5.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2019.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.3.0
toml==0.10.1
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
